Add task 6

+++ hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 {
 char c;
 int i;
-long int l;
+long long int l;
 long ln;
 float f;
 printf("Size of a char: %d byte(s)\n", (int)sizeof(c));
